정수 & 실수를 모두 받을 수 있고, 반환 타입은 정수로 고정되도록 수정

+++ src/main/java/Calculator.java
@@ -1,21 +1,21 @@
 public class Calculator {
-    public int plus(int x, int y) {
+    public int plus(double x, double y) {
         if(x > Integer.MAX_VALUE - y) throw new IllegalArgumentException("두 수의 합이 계산 가능한 범위를 벗어났습니다.");
-        return x + y;
+        return (int)(x + y);
     }
 
-    public int minus(int x, int y) {
+    public int minus(double x, double y) {
         if(x < Integer.MIN_VALUE + y) throw new IllegalArgumentException("두 수의 차가 계산 가능한 범위를 벗어났습니다.");
-        return x - y;
+        return (int)(x - y);
     }
 
-    public int multiply(int x, int y) {
+    public int multiply(double x, double y) {
         if(x > Integer.MAX_VALUE / y || x < Integer.MIN_VALUE / y) throw new IllegalArgumentException("두 수의 곱이 계산 가능한 범위를 벗어났습니다.");
-        return x * y;
+        return (int)(x * y);
     }
 
-    public int devide(int x, int y) {
+    public int devide(double x, double y) {
         if (y == 0) throw new IllegalArgumentException("0으로 나눌 수 없습니다.");
-        return x / y;
+        return (int)(x / y);
     }
 }
